+++ projects/packages/forms/src/blocks/contact-form/child-blocks.js
@@ -116,6 +116,11 @@ const FieldDefaults = {
 		to: [
 			{
 				type: 'block',
+				blocks: [ 'jetpack/field-number' ],
+				transform: attributes => createBlock( 'jetpack/field-number', attributes ),
+			},
+			{
+				type: 'block',
 				blocks: [ 'jetpack/field-text' ],
 				transform: attributes => createBlock( 'jetpack/field-text', attributes ),
 			},
@@ -376,6 +381,28 @@ export const childBlocks = [
 					type: 'string',
 					default: __( 'Text', 'jetpack-forms' ),
 					role: 'content',
+				},
+			},
+		},
+	},
+	{
+		name: 'field-number',
+		settings: {
+			...FieldDefaults,
+			title: __( 'Number Input Field', 'jetpack-forms' ),
+			description: __( 'Collect numbers from site visitors.', 'jetpack-forms' ),
+			icon: renderMaterialIcon(
+				<Path
+					fill={ getIconColor() }
+					d="M12 7H4V8.5H12V7ZM19.75 17.25V10.75H4.25V17.25H19.75ZM5.75 15.75V12.25H18.25V15.75H5.75Z"
+				/>
+			),
+			edit: editField( 'number' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: __( 'Number', 'jetpack-forms' ),
 				},
 			},
 		},