+++ projects/packages/status/src/class-host.php
@@ -235,9 +235,13 @@ class Host {
 	/**
 	 * Returns a guess of the hosting provider for the current site based on various checks.
 	 *
+	 * @since $$next-version$$ Added $guess parameter.
+	 *
+	 * @param bool $guess Whether to guess the hosting provider.
+	 *
 	 * @return string
 	 */
-	public function get_known_host_guess() {
+	public function get_known_host_guess( $guess = true ) {
 		$host = Cache::get( 'host_guess' );
 
 		if ( null !== $host ) {
@@ -267,9 +271,10 @@ class Host {
 				break;
 		}
 
-		// Second, let's check if we can recognize provider by nameservers:
+		// Second, let's check if we can recognize provider by nameservers.
+		// Only do this if we're asked to guess.
 		$domain = isset( $_SERVER['SERVER_NAME'] ) ? sanitize_text_field( wp_unslash( $_SERVER['SERVER_NAME'] ) ) : '';
-		if ( $provider === 'unknown' && ! empty( $domain ) ) {
+		if ( $provider === 'unknown' && ! empty( $domain ) && $guess ) {
 			$provider = $this->get_hosting_provider_by_nameserver( $domain );
 		}