+++ projects/packages/explat/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.7] - 2025-02-17
+### Changed
+- Update dependencies.
+
 ## [0.2.6] - 2025-02-10
 ### Changed
 - Updated package dependencies. [#41491]
@@ -103,6 +107,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ExPlat: add condition to prevent fetching the experiment assignment if there's not anon id (meaning that Tracks is likely disabled) [#38327]
 - Updated package dependencies. [#38132]
 
+[0.2.7]: https://github.com/Automattic/jetpack-explat/compare/v0.2.6...v0.2.7
 [0.2.6]: https://github.com/Automattic/jetpack-explat/compare/v0.2.5...v0.2.6
 [0.2.5]: https://github.com/Automattic/jetpack-explat/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/Automattic/jetpack-explat/compare/v0.2.3...v0.2.4